refactor: Constant 클래스 내부의 안내 메시지 상수를 InformationMessage 클래스로 분리
